# openapi.yaml
openapi: 3.0.0
info:
  title: Marketplace API
  description: API documentation for the Marketplace backend
  version: 1.0.0
servers:
  # - url: https://api.marketplace.com
  #   description: Production Server
  - url: http://localhost:8000
    description: Local Development Server
    x-cors: true
security:
  - BearerAuth: []
paths:
...
  /addresses:
    get:
      summary: Get user addresses
      description: Retrieves all saved addresses for the authenticated user.
      tags:
        - Addresses
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Successfully retrieved addresses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Address"
        "401":
          description: Unauthorized - Missing or invalid token
        "500":
          description: Internal server error
    put:
      summary: Update an address
      description: Updates a saved address for the authenticated user.
      tags:
        - Addresses
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Address"
      responses:
        "200":
          description: Address successfully updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Address"
        "400":
          description: Invalid request payload
        "401":
          description: Unauthorized - Missing or invalid token
        "500":
          description: Internal server error
    post:
      summary: Add a new address
      description: Saves a new address for the authenticated user.
      tags:
        - Addresses
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                addressee:
                  type: string
                  example: "John Doe"
                address_line1:
                  type: string
                  example: "123 Main St"
                address_line2:
                  type: string
                  nullable: true
                  example: "Apt 4B"
                city:
                  type: string
                  example: "New York"
                state_code:
                  type: string
                  example: "NY"
                postal_code:
                  type: string
                  example: "10001"
                phone:
                  type: string
                  nullable: true
                  example: "+1-555-123-4567"
      responses:
        "201":
          description: Address successfully added
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    example: "abc123"
                  addressee:
                    type: string
                    example: "John Doe"
                  address_line1:
                    type: string
                    example: "123 Main St"
                  address_line2:
                    type: string
                    nullable: true
                    example: "Apt 4B"
                  city:
                    type: string
                    example: "New York"
                  state_code:
                    type: string
                    example: "NY"
                  postal_code:
                    type: string
                    example: "10001"
                  phone:
                    type: string
                    nullable: true
                    example: "+1-555-123-4567"
                  created_at:
                    type: string
                    format: date-time
                    example: "2025-03-04T12:00:00Z"
                  updated_at:
                    type: string
                    format: date-time
                    example: "2025-03-04T12:00:00Z"
        "400":
          description: Invalid request payload
        "401":
          description: Unauthorized - Missing or invalid token
        "500":
          description: Internal server error
  /addresses/{id}:
    delete:
      summary: Remove an address
      description: Deletes a saved address for the authenticated user.
      tags:
        - Addresses
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Address successfully removed
        "400":
          description: Invalid address ID
        "401":
          description: Unauthorized - Missing or invalid token
        "500":
          description: Internal server error
...
  /products:
    post:
      summary: Create a new product
      description: Adds a new product to the catalog.
      tags:
        - Products
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: "Nordic Lounge Chair"
                price:
                  type: number
                  example: 29999
                description:
                  type: string
                  example: "A stylish and comfortable lounge chair with minimalist design."
      responses:
        "201":
          description: Product successfully created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "400":
          description: Invalid request payload
        "401":
          description: Unauthorized - Admin access required
        "500":
          description: Internal server error

    get:
      summary: Get all products
      description: Retrieves a paginated list of products with optional filtering and sorting.
      tags:
        - Products
      parameters:
        - name: category
          in: query
          description: Filter products by category ID
          schema:
            type: string
        - name: sort
          in: query
          description: Sort products by price (price_asc, price_desc)
          schema:
            type: string
            enum: [price_asc, price_desc]
        - name: in_stock
          in: query
          description: Filter products that are in stock only (true/false)
          schema:
            type: boolean
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 25
      responses:
        "200":
          description: Successfully retrieved products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
        "500":
          description: Internal server error
  /products/{id}:
    get:
      summary: Get product by ID
      description: Retrieves a single product by its ID.
      tags:
        - Products
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            example: "123"
      responses:
        "200":
          description: Successfully retrieved product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Product not found
        "500":
          description: Internal server error
  /orders:
    post:
      summary: Create a new order
      description: Creates a new order using the provided address and current cart items, returning a Stripe PaymentIntent.
      tags:
        - Orders
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                address_id:
                  type: string
                  example: "address-789"
      responses:
        "200":
          description: PaymentIntent created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentIntent"
        "400":
          description: Invalid request payload
        "401":
          description: Unauthorized - Missing or invalid token
        "500":
          description: Internal server error
    get:
      summary: Retrieve user orders
      description: Retrieves all orders for the authenticated user.
      tags:
        - Orders
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 25
      responses:
        "200":
          description: Successfully retrieved orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
        "401":
          description: Unauthorized - Missing or invalid token
        "500":
          description: Internal server error
  /orders/events:
    post:
      summary: Process Stripe events for orders
      description: Processes Stripe events related to orders.
      tags:
        - Orders
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StripeEvent"
      responses:
        "200":
          description: Event processed successfully
        "400":
          description: Invalid request payload
        "500":
          description: Internal server error
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Address:
      type: object
      properties:
        id:
          type: string
          example: "abc123"
        addressee:
          type: string
          example: "John Doe"
        address_line1:
          type: string
          example: "123 Main St"
        address_line2:
          type: string
          nullable: true
          example: "Apt 4B"
        city:
          type: string
          example: "New York"
        state_code:
          type: string
          example: "NY"
        postal_code:
          type: string
          example: "10001"
        phone:
          type: string
          nullable: true
          example: "+1-555-123-4567"
        created_at:
          type: string
          format: date-time
          example: "2025-03-04T12:00:00Z"
        updated_at:
          type: string
          format: date-time
          example: "2025-03-04T12:00:00Z"
...
    Product:
      type: object
      properties:
        id:
          type: string
          example: "123"
        name:
          type: string
          example: "Nordic Lounge Chair"
        price:
          type: integer
          example: 29999
          description: "Price in cents (e.g., 29999 represents $299.99)"
        description:
          type: string
          example: "A stylish and comfortable lounge chair with minimalist design."
        images:
          type: array
          items:
            $ref: "#/components/schemas/Image"
        created_at:
          type: string
          format: date-time
          example: "2025-03-04T12:00:00Z"
        updated_at:
          type: string
          format: date-time
          example: "2025-03-04T12:00:00Z"
    Image:
      type: object
      properties:
        id:
          type: string
          example: "image-123"
        product_id:
          type: string
          example: "123"
        image_url:
          type: string
          format: uri
          example: "https://example.com/image.jpg"
        animated:
          type: boolean
          example: false
        display_order:
          type: integer
          example: 0
        alt_text:
          type: string
          nullable: true
          example: "A modern chair with a wooden frame."
    Order:
      type: object
      properties:
        id:
          type: string
          example: "order-123"
        address_id:
          type: string
          example: "address-789"
        currency:
          type: string
          example: "usd"
        amount:
          type: integer
          example: 50000
        tax_amount:
          type: integer
          example: 4000
        shipping_amount:
          type: integer
          example: 1500
        total_amount:
          type: integer
          example: 55500
        status:
          type: string
          example: "pending"
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        created_at:
          type: string
          format: date-time
          example: "2025-03-04T12:00:00Z"
        updated_at:
          type: string
          format: date-time
          example: "2025-03-04T12:00:00Z"
    OrderItem:
      type: object
      properties:
        product_id:
          type: string
          example: "123"
        description:
          type: string
          example: "Product description"
        thumbnail:
          type: string
          example: "https://example.com/image.jpg"
        quantity:
          type: integer
          example: 2
        unit_price:
          type: integer
          example: 1000
    StripeEvent:
      type: object
      properties:
        id:
          type: string
          example: "evt_1J2YfRA12D3f5Gx"
        type:
          type: string
          example: "payment_intent.succeeded"
        data:
          $ref: "#/components/schemas/StripeEventData"
        livemode:
          type: boolean
          example: false
        created:
          type: integer
          example: 1710351234
    StripeEventData:
      type: object
      properties:
        object:
          $ref: "#/components/schemas/StripePaymentIntent"
    StripePaymentIntent:
      type: object
      properties:
        id:
          type: string
          example: "pi_3J2WfRA12D3f5Gx"
        status:
          type: string
          example: "succeeded"
        amount:
          type: integer
          example: 5000
        client_secret:
          type: string
          example: "secret_123"
        currency:
          type: string
          example: "usd"
    PaymentIntent:
      type: object
      properties:
        id:
          type: string
          example: "pi_3J2WfRA12D3f5Gx"
        amount:
          type: integer
          example: 5000
        currency:
          type: string
          example: "usd"
        status:
          type: string
          example: "requires_payment_method"
        client_secret:
          type: string
          example: "secret_123"
        error:
          type: string
          nullable: true
          example: "Error message if any"